pyproject.toml
[tool.poetry]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = ["Tacoman99 <jorgepadres789@gmail.com>"]
readme = "README.md"
packages = [{include = "ai_agents", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
crewai = {extras = ["tools"], version = "^0.83.0"}
pydantic = "^2.10.2"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
googlemaps = "^4.10.0"
requests = "^2.32.3"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
